Hide private fields; rename reset to restart.
Rather than storing the timer linked list on the timer objects, we now store a
map of task-by-id and reference the task by timer.id. This is similar to how
setTimeout and setInterval work (although we still provide a convenient timer
object with methods rather than just a long identifier).

The reset method has been renamed to restart, and you can now restart a timer
that has been stopped. If the timer was stopped, this is equivalent to creating
a new timer in that the task is put at the end of the queue.

Also, fix instanceof timer.