Add comment regarding delay calculation.
In the common case, we assume that the clock has not advanced significantly
since we woke when scheduing the next sleep timer. Since the clock can only move
forward, we only need to check the clock if we actually schedule a timer, and
not if we’ll use an animation frame.